Create README.md
It is a front-end for a hotel booking site. The site is developed using HTML/CSS and javaScript. There are total 7 subpages of the website. The slider was made on each subpage with a slightly different technique, including swiper library,  which allows creating  own pagination, navigation buttons, parallax effects and many more.
You can book a selected room or table in restaurant by filling in a simple form. You are able to find the location by using a google maps.